// src/lib/params.ts
import { MY_EDUCATIONS } from './educations.params';
import MY_EXPERIENCES from './experiences.params';
import MY_PROJECTS from './projects.params';
import MY_SKILLS from './skills.params';
import {
	Platform,
	type HomePageParams,
	type ProjectPageParams,
	type ExperiencePageParams,
	type SkillsPageParams,
	type ResumePageParams,
	type SearchPageParams,
	type EducationPageParams
} from './types';
import { Icons } from './utils';

export const TITLE_SUFFIX = 'Lim Meng Shin';

export const NavBar = {
	home: 'Home',
	personal: 'Projects',
	career: 'Experiences',
	resume: 'Resumé',
	skills: 'Skills',
	Education: 'Education'
};

export const getPlatfromIcon = (platform: Platform): Icons => {
	switch (platform) {
		case Platform.Linkedin:
			return Icons.LinkedIn;
		case Platform.GitHub:
			return Icons.GitHub;
	}
};

export const HOME: HomePageParams = {
	title: 'Home',
	name: 'Lim Meng Shin',
	description:
		'A driven tech enthusiast with a passion for turning ideas into reality. I love building projects to explore new technologies and hone my coding skills.',
	links: [
		{
			platform: Platform.Linkedin,
			link: 'https://www.linkedin.com/in/limmengshin/'
		},
		{
			platform: Platform.GitHub,
			link: 'https://github.com/LimMengShin'
		},
	]
};

export const PROJECTS: ProjectPageParams = {
	title: 'Projects',
	items: MY_PROJECTS
};

export const EXPERIENCES: ExperiencePageParams = {
	title: 'Experiences',
	items: MY_EXPERIENCES
};

export const SKILLS: SkillsPageParams = {
	title: 'Skills',
	items: MY_SKILLS
};

export const RESUME: ResumePageParams = {
	title: 'Resumé',
	item: 'resumé.pdf'
};

export const SEARCH: SearchPageParams = {
	title: 'Search'
};

export const EDUCATION: EducationPageParams = {
	items: MY_EDUCATIONS,
	title: 'Education'
};
